Prop Respawning
Added a script to respawn props when they are effected by hazards. Script uses a public method rather than it's own collision detection to respawn props. Props need a rigidbody in order to be respawned in this way. The prop is not reinstantiated, rather it has its velocity set to 0, moved to the respawn location and given a force impulse based on input in editor.